+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 * Moved deprecated methods to a separate module `DeprecatedClientMethods`.
 * Badgeville::Reward enhanced with:
   * `find_by_player` to get all the rewards for a player (reads through pagination)
+* Badgeville::RewardDefinition implemented with:
+  * `find_by_site` to get all the reward definitions for a site (returned as Badgeville::Reward)
 
 0.1.1 (May 7, 2012)
 ===================